refactor FieldPlans
= add getFieldDef to SuperSchema
= remove unnecessary member property
