+++ src/stitch/SubFieldPlan.ts
@@ -2,22 +2,11 @@ import type {
   FieldNode,
   FragmentSpreadNode,
   GraphQLCompositeType,
-  GraphQLField,
   GraphQLObjectType,
   InlineFragmentNode,
   SelectionNode,
 } from 'graphql';
-import {
-  getNamedType,
-  isAbstractType,
-  isCompositeType,
-  isInterfaceType,
-  isObjectType,
-  Kind,
-  SchemaMetaFieldDef,
-  TypeMetaFieldDef,
-  TypeNameMetaFieldDef,
-} from 'graphql';
+import { getNamedType, isAbstractType, isCompositeType, Kind } from 'graphql';
 
 import type { ObjMap } from '../types/ObjMap.js';
 
@@ -27,14 +16,18 @@ import { invariant } from '../utilities/invariant.js';
 
 import type { FieldPlan } from './FieldPlan.js';
 import { createFieldPlan } from './FieldPlan.js';
-import type { OperationContext, Subschema } from './SuperSchema.js';
+import type {
+  OperationContext,
+  Subschema,
+  SuperSchema,
+} from './SuperSchema.js';
 
 /**
  * @internal
  */
 export class SubFieldPlan {
   operationContext: OperationContext;
-  parentType: GraphQLCompositeType;
+  superSchema: SuperSchema;
   ownSelections: ReadonlyArray<SelectionNode>;
   otherSelections: ReadonlyArray<SelectionNode>;
   fieldPlans: Map<GraphQLObjectType, FieldPlan>;
@@ -49,13 +42,13 @@ export class SubFieldPlan {
     subschema: Subschema,
   ) {
     this.operationContext = operationContext;
-    this.parentType = parentType;
+    this.superSchema = operationContext.superSchema;
     this.visitedFragments = new Set();
     this.subschema = subschema;
     this.subFieldPlans = Object.create(null);
 
     const { ownSelections, otherSelections } = this._processSelections(
-      this.parentType,
+      parentType,
       selections,
     );
     this.ownSelections = ownSelections;
@@ -193,7 +186,7 @@ export class SubFieldPlan {
     }
 
     const fieldName = field.name.value;
-    const fieldDef = this._getFieldDef(parentType, fieldName);
+    const fieldDef = this.superSchema.getFieldDef(parentType, fieldName);
 
     if (!fieldDef) {
       return;
@@ -226,40 +219,6 @@ export class SubFieldPlan {
           selections: subFieldPlan.otherSelections,
         },
       });
-    }
-  }
-
-  _getFieldDef(
-    parentType: GraphQLCompositeType,
-    fieldName: string,
-  ): GraphQLField<any, any> | undefined {
-    if (fieldName === '__typename') {
-      return TypeNameMetaFieldDef;
-    }
-
-    invariant(
-      isObjectType(parentType) || isInterfaceType(parentType),
-      `Invalid parent type ${inspect(parentType)}.`,
-    );
-
-    const fields = parentType.getFields();
-
-    const field = fields[fieldName];
-
-    if (field !== undefined) {
-      return field;
-    }
-
-    if (
-      parentType ===
-      this.operationContext.superSchema.mergedSchema.getQueryType()
-    ) {
-      switch (fieldName) {
-        case SchemaMetaFieldDef.name:
-          return SchemaMetaFieldDef;
-        case TypeMetaFieldDef.name:
-          return TypeMetaFieldDef;
-      }
     }
   }
